requirements.txt
asgiref==3.7.2
attrs==24.2.0
backports.zoneinfo==0.2.1
Django==3.2.25
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.28.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
pkgutil-resolve-name==1.3.10
PyJWT==2.8.0
pyrsistent==0.19.3
pytz==2025.2
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
zipp==3.15.0
